refactor(add): split add.rs into spec/manifest/global submodules
`crates/aube/src/commands/add.rs` had grown to 2901 lines mixing CLI
policy, package spec parsing, manifest mutation, catalog/workspace
protocol behavior, supply-chain checks, no-save restore, and global
install. Split along those seams, matching the `commands/install/`
layout:

- `add/spec.rs`: `ParsedPkgSpec` + parser (jsr/git/local/alias variants)
  and its 38 unit tests
- `add/manifest.rs`: `AddManifestOptions`, `update_manifest_for_add`,
  the `apply_*_to_manifest` writers, `decide_save_catalog`, workspace
  version discovery
- `add/global.rs`: `run_global` + `run_global_inner`
- `add/mod.rs`: `AddArgs`, `run`, `run_filtered`, supply-chain helpers,
  allow/deny-build plumbing, `NoSaveSnapshot`, `lockfile_path_for_project`

Cross-module items use `pub(super)`. Public surface (`AddArgs`, `run`)
is unchanged.
